removing specific species lists
