+++ src/components/Mechanism/GasSpeciesTab.js
@@ -1,18 +1,34 @@
 import React from "react";
 import { connect } from "react-redux";
-import GasSpeciesList from "./GasSpeciesList.js";
+
+import SpeciesList from "./SpeciesList";
+import AddSpecies from "./AddSpecies";
+import Species from "./Species";
 import SpeciesTab from "./SpeciesTab.js";
 import GasPropertyList from "./GasPropertyList.js";
 import SpeciesDetail from "./SpeciesDetail.js";
+import { getMechanism } from "../../redux/selectors";
 
-function GasSpeciesTab() {
+function GasSpeciesTab(props) {
   return (
     <SpeciesTab
       speciesType="gas"
-      SpeciesList={GasSpeciesList}
-      SpeciesDetail={(props) => <SpeciesDetail {...props} propertyListComponent={GasPropertyList}/>}
+      SpeciesList={
+        (specListProps) => <SpeciesList 
+          {...specListProps}
+          species={props.gasSpecies} 
+          addSpeciesComponent={() => <AddSpecies type="gas"/>} 
+          speciesComponent={(specCompProps) => <Species {...specCompProps} type="gas"/>} 
+        />
+      }
+      SpeciesDetail={(specDetailProps) => <SpeciesDetail {...specDetailProps} propertyListComponent={GasPropertyList}/>}
     />
   );
 }
 
-export default connect()(GasSpeciesTab);
+const mapStateToProps = (state) => {
+  const mechanism = getMechanism(state);
+  return mechanism;
+};
+
+export default connect(mapStateToProps)(GasSpeciesTab);
