+++ src/components/Mechanism/index.js
@@ -1,12 +1,10 @@
 export { default as AddAerosolProperty } from './AddAerosolProperty'
 export { default as AerosolPropertyList } from './AerosolPropertyList'
-export { default as AerosolSpeciesList } from './AerosolSpeciesList'
 export { default as AerosolSpeciesTab } from './AerosolSpeciesTab'
 export { default as AddProduct } from './AddProduct'
 export { default as AddProperty } from './AddProperty'
 export { default as AddReactant } from './AddReactant'
 export { default as AddReaction } from './AddReaction'
-export { default as GasSpeciesList } from './GasSpeciesList'
 export { default as GasSpeciesTab } from './GasSpeciesTab'
 export { default as Product } from './Product'
 export { default as ProductList } from './ProductList'
